+++ tests/test_prg_engine_runtime_surface_functions_buffering_main.cpp
@@ -7,7 +7,8 @@ int main()
     using namespace copperfin::test_support;
     test_curval_evaluates_on_disk_record();
     test_curval_uses_verified_post_commit_session_image();
-    test_curval_verified_commit_overlay_does_not_survive_cursor_reopen();
+    test_curval_verified_commit_overlay_survives_cursor_reopen();
+    test_curval_verified_commit_overlay_is_visible_to_a_second_concurrent_cursor();
     test_getfldstate_tracks_buffered_mutation_state();
     test_setfldstate_assigns_buffered_mutation_state();
     test_oldval_evaluates_buffered_original_record();